refactor: migrate manual factory to Spring DI and add JPA repositories

+++ src/main/java/com/carlosedolv/contractflow/services/contracts/FeeCalculator.java
@@ -1,6 +1,8 @@
 package com.carlosedolv.contractflow.services.contracts;
 
+import java.math.BigDecimal;
+
 public interface FeeCalculator {
-	Double applyMonthlyFee(Double base, int month);
-	Double applyPaymentFee(Double amount);
+    BigDecimal applyMonthlyFee(BigDecimal base, int month);
+    BigDecimal applyPaymentFee(BigDecimal amount);
 }
